+++ client/src/components/auth/Login.css
@@ -1,12 +1,8 @@
-#login-page {
-  height: 100vh;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
 #form-box {
   box-shadow: 0px 0px 15px -5px black;
+  width: 40%;
+  min-width: 400px;
+  min-height: 450px;
 }
 
 #loginLink {
@@ -14,5 +10,5 @@
   float: left;
 }
 #loginButton {
-  float: left;
+  float: right;
 }
