+++ client/src/components/auth/LoginPage.jsx
@@ -6,11 +6,12 @@ import React from 'react';
 import AuthBackground from './authComponents/AuthBackground';
 import RegisterButton from './authComponents/RegisterButton';
 import AuthLayout from './AuthLayout';
+import LoginForm from './LoginForm';
 
 const LoginPage = () => (
   <>
     <AuthLayout button={RegisterButton()}>
-      <AuthBackground />
+      <AuthBackground form={LoginForm()} />
     </AuthLayout>
   </>
 );
